+++ crates/nu-explore/src/pager/mod.rs
@@ -484,14 +484,14 @@ fn set_cursor_cmd_bar(f: &mut Frame, area: Rect, pager: &Pager) {
         let next_pos = (pager.cmd_buf.buf_cmd2.len() + 1) as u16;
         // 1 skips a ':' char
         if next_pos < area.width {
-            f.set_cursor(next_pos as u16, area.height - 1);
+            f.set_cursor(next_pos, area.height - 1);
         }
     } else if pager.search_buf.is_search_input {
         // todo: deal with a situation where we exceed the bar width
         let next_pos = (pager.search_buf.buf_cmd_input.len() + 1) as u16;
         // 1 skips a ':' char
         if next_pos < area.width {
-            f.set_cursor(next_pos as u16, area.height - 1);
+            f.set_cursor(next_pos, area.height - 1);
         }
     }
 }
