+++ crates/nu-explore/src/views/interactive.rs
@@ -112,10 +112,10 @@ impl View for InteractiveView<'_> {
         f.render_widget(cmd_input, cmd_input_area);
 
         if !self.view_mode {
-            let cur_w = area.x + 1 + 1 + 1 + max_cmd_len as u16;
+            let cur_w = area.x + 1 + 1 + 1 + max_cmd_len;
             let cur_w_max = area.x + 1 + 1 + 1 + area.width - 2 - 1 - 1 - 1 - 1;
             if cur_w < cur_w_max {
-                f.set_cursor(area.x + 1 + 1 + 1 + max_cmd_len as u16, area.y + 1);
+                f.set_cursor(area.x + 1 + 1 + 1 + max_cmd_len, area.y + 1);
             }
         }
